feat: add rebuild from existing function (#84)
It's useful to be able to rebuild a Compressor from an existing symbol
table. For example, if you want to perform incremental compression, or
compress a value to pushdown comparisons over the encoded data.

Going through `CompressorBuilder`, calling `insert` for each symbol and
`build`ing may seem appropriate, but you're not guaranteed to get the
same codes order afterward! This is because of the `finalize` step, many
of the 2-byte codes will actually get reordered. This is an optimization
we port from the C++ code, and allows us to skip a hashtable lookup
during compression for many inputs.

To avoid the ambiguity, we provide a new `Compressor::rebuild_from` to
rebuild from a slice of symbols and lens. This repopulates the
`codes_two_byte` and hash table that are needed at compress time, and
does so without the finalize step that occurs when going through the
CompressorBuilder.

I've also gone and updated the fuzzer to check both a freshly built and
a rebuilt compressor. It caught several bugs as I was going through
cleaning up this PR 😅 I let the fuzzer run for 30 minutes on my laptop
and it didn't find anything.